Developed Object-Oriented Database System v4
