+++ src/lib/server/database.ts
@@ -1,7 +1,7 @@
 // THIS IS THE UPDATED VERSION OF database.ts
 // version 3.0.0
 // author: @tinykishore
-import {Db, MongoClient, ObjectId} from "mongodb";
+import {Db, MongoClient} from "mongodb";
 import {MONGO_DATABASE, MONGO_URL} from "$env/static/private";
 import consoleLog, {LEVEL} from "$lib/server/log";
 
@@ -14,16 +14,7 @@ let databaseConnection: Db | null = null;
 /**
  * An object that stores various collections used for database operations.
  */
-let collections: any = {
-    "user-account": null,
-    "farm-info": null,
-    "farm-products": null,
-    "forum": null,
-    "comment": null,
-    "news": null,
-    "order": null,
-    "animal-health-details": null,
-}
+let collections: any = {}
 
 /**
  * The MongoDB client used to connect to the MongoDB server.
@@ -43,17 +34,19 @@ export const initializeDatabaseConnection = async (): Promise<void> => {
         await client.connect();
         consoleLog("DATABASE LOG: Connected to MongoDB Server", LEVEL.OK)
         databaseConnection = client.db(MONGO_DATABASE);
-        collections["user-account"] = databaseConnection.collection("user-account");
-        collections["farm-info"] = databaseConnection.collection("farm-info");
-        collections["farm-products"] = databaseConnection.collection("farm-products");
-        collections["forum"] = databaseConnection.collection("forum");
-        collections["comment"] = databaseConnection.collection("comment");
-        collections["news"] = databaseConnection.collection("news");
-        collections["order"] = databaseConnection.collection("order");
-        collections["animal-health-details"] = databaseConnection.collection("animal-health-details");
 
+        // Dynamically fetch all collections in the database
+        consoleLog("DATABASE LOG: Fetching collections...", LEVEL.INFO)
+        const collectionsList = await databaseConnection.listCollections().toArray();
 
-        // More to add here
+        // Create collections dynamically
+        consoleLog("DATABASE LOG: Creating collections...", LEVEL.INFO)
+        for (const collectionInfo of collectionsList) {
+            collections[collectionInfo.name] = databaseConnection.collection(collectionInfo.name);
+            // consoleLog(`DATABASE LOG: Created collection {` + collectionInfo.name + `}`, LEVEL.INFO)
+        }
+        consoleLog("DATABASE LOG: Collections created successfully", LEVEL.OK)
+        console.log(collections)
     } catch (error: any) {
         consoleLog(`DATABASE ERROR: ${error.message}`, LEVEL.ERROR);
         databaseConnection = null;
@@ -70,6 +63,7 @@ export const isDatabaseConnectionInitialized = (): boolean => {
 }
 
 /**
+ * noinspection JSUnusedLocalSymbols
  * Terminates the database connection, closing the connection to the MongoDB server.
  *
  * @returns A Promise that resolves when the connection is successfully closed.
@@ -86,380 +80,10 @@ export const terminateDatabaseConnection = async (): Promise<void> => {
 /**
  * A class for interacting with the database.
  */
-export class Database {
-    /**
-     * Retrieves a user from the database by their email.
-     *
-     * @param email - The email of the user to look up.
-     * @returns A Promise that resolves to the user object or null if not found.
-     */
-    public static async getUserByEmail(email: string): Promise<any> {
-        consoleLog("DATABASE LOG: Getting user by email...", LEVEL.OK)
-        return await collections["user-account"].findOne({"credentials.email": email});
-    }
+export default class Database {
 
-    /**
-     * Retrieves a user from the database by their username.
-     *
-     * @param username - The username of the user to look up.
-     * @returns A Promise that resolves to the user object or null if not found.
-     */
-    public static async getUserByUsername(username: string): Promise<any> {
-        consoleLog("DATABASE LOG: Getting user by username...", LEVEL.OK)
-        return await collections["user-account"].findOne({"credentials.username": username});
-    }
-
-    /**
-     * Retrieves a user from the database by their unique ObjectID.
-     *
-     * @param _id - The ObjectID of the user to look up.
-     * @returns A Promise that resolves to the user object or null if not found.
-     */
-    public static async getUserByObjectID(_id: ObjectId): Promise<any> {
-        consoleLog("DATABASE LOG: Getting user by ObjectID...", LEVEL.OK)
-        return await collections["user-account"].findOne({_id: _id});
-    }
-
-    /**
-     * Retrieves cached user information from the database by their unique ObjectID.
-     *
-     * @param _id - The ObjectID of the user to retrieve cache information for.
-     * @returns A Promise that resolves to a TypeUserCache object.
-     */
-    public static async getUserCache(_id: ObjectId) {
-        consoleLog(`DATABASE LOG: Getting {` + _id + `} cache information...`, LEVEL.OK)
-        const result = await collections["user-account"].findOne({_id: _id});
-        return {
-            _id: result._id,
-            full_name: result.full_name,
-            email: result.credentials.email,
-            username: result.credentials.username,
-            profile_picture: result.profile_picture,
-            user_role: result.role,
-        };
-    }
-
-    /**
-     * Retrieves user credentials from the database by their unique ObjectID.
-     *
-     * @param _id - The ObjectID of the user to retrieve credentials for.
-     * @returns A Promise that resolves to a Credentials object.
-     */
-    public static async getUserCredentialsByObjectID(_id: ObjectId): Promise<Credentials> {
-        consoleLog("DATABASE LOG: Getting user credentials by ObjectID...", LEVEL.OK)
-        const result = await collections["user-account"].findOne({_id: _id});
-        return result.credentials;
-    }
-
-    /**
-     * Inserts a reset password token for a user by their unique ObjectID.
-     *
-     * @param email - The ObjectID of the user for whom the reset token is inserted.
-     * @param token - The reset password token to be inserted.
-     * @returns A boolean indicating whether the reset token insertion was successful.
-     */
-    public static async insertResetPasswordToken(email: string, token: string): Promise<boolean> {
-        const result = await collections["user-account"].updateOne(
-            {"credentials.email": email},
-            {$set: {"credentials.password_reset_token": token}}
-        );
-        if (result.modifiedCount === 1) {
-            consoleLog("DATABASE LOG: Reset token inserted successfully", LEVEL.OK);
-            return true;
-        }
-        consoleLog("DATABASE LOG: Reset token insertion failed", LEVEL.ERROR);
-        return false;
-    }
-
-    public static async getResetPasswordToken(token: string) {
-        consoleLog("DATABASE LOG: Getting reset password token...", LEVEL.OK)
-        const result = await collections["user-account"].findOne({"credentials.password_reset_token": token});
-        if (result !== null) {
-            consoleLog("DATABASE LOG: Reset token found", LEVEL.OK);
-            return {
-                _id: result._id,
-                success: true
-            }
-        }
-        consoleLog("DATABASE LOG: Reset token not found", LEVEL.ERROR);
-        return {success: false}
-    }
-
-    public static async deleteResetPasswordToken(_id: ObjectId) {
-        consoleLog("DATABASE LOG: Deleting reset password token...", LEVEL.OK)
-        const result = await collections["user-account"].updateOne(
-            {_id: _id},
-            {$unset: {"credentials.password_reset_token": ""}}
-        );
-        if (result.modifiedCount === 1) {
-            consoleLog("DATABASE LOG: Reset token deleted successfully", LEVEL.OK);
-            return true;
-        }
-
-        consoleLog("DATABASE LOG: Reset token deletion failed", LEVEL.ERROR);
-        return false;
-
-    }
-
-    public static async updatePassword(_id: ObjectId, password: string) {
-        consoleLog("DATABASE LOG: Updating password...", LEVEL.OK)
-        const result = await collections["user-account"].updateOne(
-            {_id: _id},
-            {$set: {"credentials.password_hash": password}}
-        );
-        if (result.modifiedCount === 1) {
-            consoleLog("DATABASE LOG: Password updated successfully", LEVEL.OK);
-            return true;
-        }
-        consoleLog("DATABASE LOG: Password update failed", LEVEL.ERROR);
-        return false;
-    }
-
-    public static async updateGoogleID(email: string, google_id: string) {
-        consoleLog("DATABASE LOG: Updating Google ID...", LEVEL.OK)
-        const result = await collections["user-account"].updateOne(
-            {"credentials.email": email},
-            {$set: {"credentials.google_id": google_id}}
-        );
-        if (result.modifiedCount === 1) {
-            consoleLog("DATABASE LOG: Google ID updated successfully", LEVEL.OK);
-            return true;
-        }
-        consoleLog("DATABASE LOG: Google ID update failed", LEVEL.ERROR);
-        return false;
-    }
-
-    public static async insertUser(userObject: UserObject): Promise<any> {
-        consoleLog("DATABASE LOG: Inserting user...", LEVEL.OK)
-        return await collections["user-account"].insertOne(userObject);
-    }
-
-    public static async completeGoogleSignUp(_id: ObjectId, username: string, password_hash: string) {
-        // insert username and password_hash into database for user with _id
-        consoleLog("DATABASE LOG: Completing Google Sign Up...", LEVEL.OK)
-        const result = await collections["user-account"].updateOne(
-            {_id: _id},
-            {$set: {"credentials.username": username, "credentials.password_hash": password_hash}}
-        );
-        if (result.modifiedCount === 1) {
-            consoleLog("DATABASE LOG: Google Sign Up completed successfully", LEVEL.OK);
-            return true;
-        }
-        consoleLog("DATABASE LOG: Google Sign Up failed", LEVEL.ERROR);
-        return false;
-    }
-
-    public static async updateProfilePicture(_id: ObjectId, profile_picture: string) {
-        consoleLog("DATABASE LOG: Updating profile picture...", LEVEL.OK)
-        const result = await collections["user-account"].updateOne(
-            {_id: _id},
-            {$set: {"profile_picture": profile_picture}}
-        );
-        if (result.modifiedCount === 1) {
-            consoleLog("DATABASE LOG: Profile picture updated successfully", LEVEL.OK);
-            return true;
-        }
-        consoleLog("DATABASE LOG: Profile picture update failed", LEVEL.ERROR);
-        return false;
-    }
-
-    public static async crosscheckUsernameAndObjectID(username: string, _id: ObjectId) {
-        consoleLog("DATABASE LOG: Crosschecking username and ObjectID...", LEVEL.OK)
-        const result = await collections["user-account"].findOne({"credentials.username": username});
-        if (result !== null) {
-            if (result._id.equals(_id)) {
-                consoleLog("DATABASE LOG: Username and ObjectID match", LEVEL.OK);
-                return true;
-            }
-        }
-        consoleLog("DATABASE LOG: Username and ObjectID do not match", LEVEL.ERROR);
-        return false;
-    }
-
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // END OF USER ACCOUNT DATABASE OPERATIONS. START OF FARM INFO DATABASE OPERATIONS
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // ##############################################################################################################
-
-    /**
-     * Retrieves information about all farms from the database.
-     *
-     * @returns A Promise that resolves to an array of farm information objects.
-     */
-    public static async getAllFarms(): Promise<any> {
-        consoleLog("DATABASE LOG: Getting all farms information...", LEVEL.OK)
-        return await collections["farm-info"].find({}).toArray();
-    }
-
-    /**
-     * Retrieves information about a single farm by its unique identifier (UID).
-     *
-     * @param farm_uid - The unique identifier (UID) of the farm to retrieve information for.
-     * @returns A Promise that resolves to the farm information object or null if not found.
-     */
-    public static async getOneFarm(farm_uid: string): Promise<any> {
-        consoleLog(`DATABASE LOG: Getting farm {` + farm_uid + `} information...`, LEVEL.OK)
-        return await collections["farm-info"].findOne({"uid": farm_uid});
-    }
-
-    // /**
-    //  * Retrieves information about all farm products from the database.
-    //  *
-    //  * @returns A Promise that will eventually contain the data once the implementation is completed.
-    //  */
-    // public static async getAllFarmProducts(): Promise<any> {
-    //
-    // }
-
-    /**
-     * Retrieves information about farm products associated with a single farm by its unique identifier (UID).
-     *
-     * @param farm_uid - The unique identifier (UID) of the farm to retrieve product information for.
-     * @returns A Promise that resolves to the farm products information object or null if not found.
-     */
-    public static async getOneFarmProducts(farm_uid: string): Promise<any> {
-        consoleLog(`DATABASE LOG: Getting farm products {` + farm_uid + `} information...`, LEVEL.OK)
-        return await collections["farm-products"].findOne({"uid": farm_uid});
-    }
-
-    /**
-     * Retrieves information about a single product by its unique identifier (ID).
-     *
-     * @param product_id - The unique identifier (ID) of the product to retrieve information for.
-     * @returns A Promise that resolves to the product information object or null if not found.
-     */
-    public static async getOneProduct(product_id: string): Promise<any> {
-        consoleLog(`DATABASE LOG: Getting product {` + product_id + `} information...`, LEVEL.OK)
-        const result = await collections["farm-products"].findOne({'products.id': product_id});
-        const products = result.products;
-        for (let i: number = 0; i < products.length; i++) {
-            if (products[i].id === product_id) {
-                return products[i];
-            }
-        }
-        return null;
-    }
-
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // END OF FARM INFO DATABASE OPERATIONS. START OF FORUM DATABASE OPERATIONS
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // ##############################################################################################################
-
-    /**
-     * Inserts a new post into the forum.
-     *
-     * @param post - The post to be inserted.
-     * @returns A Promise that resolves to the result of the insertion operation.
-     */
-    public static async insertPost(post: Post): Promise<any> {
-        consoleLog(`DATABASE LOG: Inserting post {` + post + `} information...`, LEVEL.OK)
-        return await collections["forum"].insertOne(post);
-    }
-
-    // /**
-    //  * Edits an existing post in the forum.
-    //  *
-    //  * @returns A Promise that will eventually contain the result of the edit operation once the implementation is completed.
-    //  */
-    // public static async editPost(): Promise<any> {
-    //     //
-    // }
-    //
-    // /**
-    //  * Deletes an existing post from the forum.
-    //  *
-    //  * @returns A Promise that will eventually contain the result of the delete operation once the implementation is completed.
-    //  */
-    // public static async deletePost(): Promise<any> {
-    //     //
-    // }
-
-    /**
-     * Retrieves information about a single post from the forum by its unique ObjectID.
-     *
-     * @param post_uid - The ObjectID of the post to retrieve information for.
-     * @returns A Promise that resolves to the post information object or null if not found.
-     */
-    public static async getOnePost(post_uid: ObjectId): Promise<any> {
-        consoleLog(`DATABASE LOG: Getting post {` + post_uid + `} information...`, LEVEL.OK)
-        return await collections["forum"].findOne({_id: post_uid});
-    }
-
-    /**
-     * Retrieves information about all posts from the forum.
-     *
-     * @returns A Promise that resolves to an array of post information objects.
-     */
-    public static async getAllPosts(): Promise<any> {
-        consoleLog("DATABASE LOG: Getting all posts...", LEVEL.OK)
-        return await collections["forum"].find({}).toArray();
-    }
-
-    /**
-     * Retrieves information about the most liked posts from the forum.
-     *
-     * @returns A Promise that resolves to an array of the most liked post information objects.
-     */
-    public static async getMostLikedPosts(): Promise<any> {
-        consoleLog("DATABASE LOG: Getting all most liked posts...", LEVEL.OK)
-        return await collections["forum"].find({}).sort({like: -1}).limit(3).toArray();
-    }
-
-    /**
-     * Inserts a reply into a post in the comment section.
-     *
-     * @param reply - The reply to be inserted.
-     * @returns A Promise that resolves to the result of the insertion operation.
-     */
-    public static async insertReplyInPost(reply: any): Promise<any> {
-        consoleLog(`DATABASE LOG: Sending reply {` + reply + `} information...`, LEVEL.OK)
-        return await collections["comment"].insertOne(reply);
-    }
-
-    public static async votePost(postObjectID: ObjectId, likerObjectID: string, alreadyLiked: boolean): Promise<any> {
-        consoleLog(`DATABASE LOG: Voting post {` + postObjectID + `} by user {` + likerObjectID + `}...`, LEVEL.OK)
-        if (!alreadyLiked) {
-            consoleLog(`DATABASE LOG: User {` + likerObjectID + `} already liked post {` + postObjectID + `}. Removing like...`, LEVEL.WARN)
-            return await collections["forum"].updateOne(
-                {_id: postObjectID},
-                {$pull: {likes: likerObjectID}}
-            );
-        } else {
-            consoleLog(`DATABASE LOG: User {` + likerObjectID + `} has not liked post {` + postObjectID + `}. Adding like...`, LEVEL.WARN)
-            return await collections["forum"].updateOne(
-                {_id: postObjectID},
-                {$addToSet: {likes: likerObjectID}}
-            );
-        }
-    }
-
-    /**
-     * Increments the view count for a post in the forum.
-     *
-     * @param _id - The ObjectID of the post for which the view count should be incremented.
-     * @returns A Promise that resolves to the result of the view count increment operation.
-     */
-    public static async incrementViewCount(_id: ObjectId): Promise<any> {
-        consoleLog("DATABASE LOG: Incrementing view count...", LEVEL.OK)
-        return await collections["forum"].updateOne(
-            {_id: _id},
-            {$inc: {viewCount: 1}}
-        );
-    }
-
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // END OF FORUM DATABASE OPERATIONS. START OF NEWS DATABASE OPERATIONS
-    // ##############################################################################################################
-    // ##############################################################################################################
-    // ##############################################################################################################
+    // Declare private static variables
+    protected static collections = collections;
 
     public static async getAllNews() {
         consoleLog("DATABASE LOG: Getting all news information...", LEVEL.OK)
