+++ src/lib/server/databaseObjects/DatabaseNews.ts
@@ -1,0 +1,6 @@
+import Database from "$lib/server/database";
+
+
+export default class DatabaseNews extends Database {
+    // Will be implemented in the future.
+}
