+++ src/routes/API/v1/farms/GetFarmsAPI/+server.ts
@@ -1,6 +1,6 @@
 import consoleLog, {LEVEL} from "$lib/server/log";
 import {verifyRequest} from "$lib/server/utility";
-import {Database} from "$lib/server/database";
+import DatabaseFarm from "$lib/server/databaseObjects/DatabaseFarm";
 
 export const GET = async ({cookies}: any) => {
     consoleLog("GetFarmsAPI REQUEST Received", LEVEL.OK);
@@ -8,7 +8,7 @@ export const GET = async ({cookies}: any) => {
     if (!verifyRequest(cookies)) {
         return new Response("Unauthorized", {status: 401});
     }
-    const allFarmsInfo = await Database.getAllFarms();
+    const allFarmsInfo = await DatabaseFarm.getAllFarms();
     if (!allFarmsInfo) {
         consoleLog("GetFarmsAPI RESPONSE: status 404", LEVEL.ERROR);
         return new Response(null, {status: 404});
